avoid recursive dependency when it is not needed in Melcor/Melgen
